backend/requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.7.0,<3.0.0
openai>=1.40.0,<2.0.0
langchain>=0.2.0,<0.3.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pandas>=2.2.0
python-dotenv>=1.0.1
langchain-openai>=0.1.12
langchain-community>=0.2.10
tiktoken>=0.7.0
sentence-transformers>=3.0.0

